# openapi.yml
...
openapi: 3.0.0
paths:
...
  /media/generate-video:
    post:
      operationId: MediaController_generateVideo
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateVideoDto"
      responses:
        "201":
          description: ""
      tags:
        - Media
  /media/jobs/{id}:
    get:
      operationId: MediaController_getMediaJob
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Media
...
  /media/icons/resolve:
    get:
      operationId: MediaController_resolveIcon
      parameters:
        - name: name
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
      summary: |-
        `mdi:rocket` -> the icon's raw SVG body.

        Insert > Stock Icon routed an Iconify pick through the image path, which
        stored the API URL as a raster `src` — a 24x24 black square with no fill
        control. An `icon` element wants the markup itself, and the resolver that
        produces it (safeFetch + script-stripping + cache) is server-side, so the
        editor asks for it here rather than reaching api.iconify.design directly
        past the frontend's connect-src.
      tags:
        - Media
  /media/video/function:
    post:
      operationId: MediaController_videoFunction
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoFunctionDto"
      responses:
        "201":
          description: ""
      tags:
        - Media
  /media/generate-video/{type}/allowed:
    get:
      operationId: MediaController_generateVideoAllowed
      parameters:
        - name: type
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Media
...
  /media/remove-background:
    post:
      operationId: MediaController_removeBackground
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemoveBackgroundDto"
      responses:
        "201":
          description: ""
      tags:
        - Media
  /media/inpaint:
    post:
      operationId: MediaController_inpaint
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InpaintImageDto"
      responses:
        "201":
          description: ""
      tags:
        - Media
  /media/upscale:
    post:
      operationId: MediaController_upscale
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpscaleImageDto"
      responses:
        "201":
          description: ""
      tags:
        - Media
  /media/image-to-image:
    post:
      operationId: MediaController_imageToImage
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageToImageDto"
      responses:
        "201":
          description: ""
      tags:
        - Media
  /media/upscale-video:
    post:
      operationId: MediaController_upscaleVideo
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpscaleVideoDto"
      responses:
        "201":
          description: ""
      tags:
        - Media
  /media/remove-video-background:
    post:
      operationId: MediaController_removeVideoBackground
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemoveVideoBackgroundDto"
      responses:
        "201":
          description: ""
      tags:
        - Media
  /media/video-to-video:
    post:
      operationId: MediaController_videoToVideo
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoToVideoDto"
      responses:
        "201":
          description: ""
      tags:
        - Media
  /media/generate-music:
    post:
      operationId: MediaController_generateMusic
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateMusicDto"
      responses:
        "201":
          description: ""
      tags:
        - Media
  /media/generate-avatar:
    post:
      operationId: MediaController_generateAvatar
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateAvatarDto"
      responses:
        "201":
          description: ""
      tags:
        - Media
  /media/generate-slide:
    post:
      operationId: MediaController_generateSlide
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateSlideDto"
      responses:
        "201":
          description: ""
      tags:
        - Media
...
  /media/fonts/{fileId}:
    delete:
      operationId: MediaController_deleteFont
      parameters:
        - name: fileId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
      tags:
        - Media
  /media/voices:
    get:
      operationId: MediaController_listVoices
      parameters:
        - name: provider
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Media
...
  /media/detect-focal-point:
    post:
      operationId: MediaController_detectFocalPoint
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DetectFocalPointDto"
      responses:
        "201":
          description: ""
      tags:
        - Media
...
info:
  title: ValidPost API
  description: OpenAPI description of the ValidPost backend (NestJS). Generated from the controllers — do not hand-edit; run `pnpm run openapi:generate`.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    api-key:
      type: apiKey
      in: header
      name: Authorization
    bearer:
      scheme: bearer
      bearerFormat: Token
      type: http
  schemas:
...
    GenerateVideoDto:
      type: object
      properties:
        prompt:
          type: string
        imageUrl:
          type: string
        type:
          type: string
        output:
          type: string
    VideoFunctionDto:
      type: object
      properties:
        identifier:
          type: string
          description: Video generator identifier.
        functionName:
          type: string
          description: Generator function/method to invoke.
        params:
          type: object
          description: Arguments passed to the generator function.
      required:
        - identifier
        - functionName
    RemoveBackgroundDto:
      type: object
      properties:
        imageUrl:
          type: string
          format: uri
      required:
        - imageUrl
    InpaintImageDto:
      type: object
      properties:
        imageUrl:
          type: string
          format: uri
        maskUrl:
          type: string
          format: uri
        prompt:
          type: string
      required:
        - imageUrl
        - maskUrl
        - prompt
    UpscaleImageDto:
      type: object
      properties:
        imageUrl:
          type: string
          format: uri
        scale:
          type: number
          minimum: 1
          maximum: 8
      required:
        - imageUrl
    ImageToImageDto:
      type: object
      properties:
        imageUrl:
          type: string
          format: uri
        prompt:
          type: string
      required:
        - imageUrl
        - prompt
    UpscaleVideoDto:
      type: object
      properties:
        videoUrl:
          type: string
          format: uri
        scale:
          type: number
          minimum: 1
          maximum: 8
      required:
        - videoUrl
    RemoveVideoBackgroundDto:
      type: object
      properties:
        videoUrl:
          type: string
          format: uri
      required:
        - videoUrl
    VideoToVideoDto:
      type: object
      properties:
        videoUrl:
          type: string
          format: uri
        prompt:
          type: string
      required:
        - videoUrl
        - prompt
    GenerateMusicDto:
      type: object
      properties:
        prompt:
          type: string
      required:
        - prompt
    GenerateAvatarDto:
      type: object
      properties:
        script:
          type: string
        imageUrl:
          type: string
          format: uri
      required:
        - script
    GenerateSlideDto:
      type: object
      properties:
        prompt:
          type: string
        imageUrls:
          type: array
          items:
            type: string
            format: uri
      required:
        - prompt
    DetectFocalPointDto:
      type: object
      properties:
        imageUrl:
          type: string
          format: uri
      required:
        - imageUrl
...